+++ docs/.vuepress/navbar/en.ts
@@ -1,0 +1,7 @@
+import { navbar } from "vuepress-theme-hope";
+
+export const enNavbarConfig = navbar([
+    { text: 'Guide', link: '/en/guide/' },
+    { text: 'NuGet', link: 'https://www.nuget.org/packages?q=freesql' },
+    { text: 'Api', link: 'https://docs.dotnet-china.com/FreeSql/api/index.html' },
+]);